requirements.txt
beautifulsoup4
requests
Pillow
img2pdf
pyinstaller
numpy
tqdm